% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods.R
\docType{methods}
\name{sig.genes}
\alias{sig.genes}
\alias{sig.genes,svplsTest-method}
\alias{sig.genes.svplsTest}
\title{Accessor for the 'sig.genes' slot of a 'svplsTest' object}
\usage{
sig.genes(object)

\S4method{sig.genes}{svplsTest}(object)

\S4method{sig.genes}{svplsTest}(object)
}
\arguments{
\item{object}{a \code{svplsTest} object}
}
\value{
A vector of the genes detected to be significantly differentially expressed
between the two specified conditions by \code{\link{svplsTest}}.
}
\description{
Accessor for the 'sig.genes' slot of a 'svplsTest' object
}
\examples{
data(sim.dat)
group = as.factor(c(rep(1, 10), rep(-1, 10)))
sv = svplsSurr(sim.dat, group)
surr = surr(sv)
fit = svplsTest(dat = sim.dat, group = group, surr = surr, test = "Wald")
sig.genes(fit)
}

