\name{sim.dat}
\alias{sim.dat}
\docType{data}
\title{
A simulated RNAseq gene expression count data affected by multiple hidden variables.
}
\description{
The dataset contains simulated raw RNAseq expression counts corresponding to 1000
genes over 20 subjects S1, S2..S20, distributed equally between two groups 1 and 2.
(S1, S2...S10 belong to group 1 and S11, S12...S20 belong to group 2). The data is affected by the unknown effects from several technical and sample-specific artefacts. The data has been created to illustrate usage of the functions in this package.
}
