% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods.R
\docType{methods}
\name{surr}
\alias{surr}
\alias{surr,svplsSurr-method}
\alias{surr.svplsSurr}
\title{Accessor for the 'surr' slot of a 'svplsSurr' object}
\usage{
surr(object)

\S4method{surr}{svplsSurr}(object)

\S4method{surr}{svplsSurr}(object)
}
\arguments{
\item{object}{a \code{svplsSurr} object}
}
\value{
A matrix of the significant surrogate variables obtained from
\code{\link{svplsSurr}}
}
\description{
Accessor for the 'surr' slot of a 'svplsSurr' object
}
\examples{
data(sim.dat)
group = as.factor(c(rep(1, 10), rep(-1, 10)))
sv = svplsSurr(sim.dat, group)
surr(sv)
}

