% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{svplsSurr-class}
\alias{svplsSurr-class}
\title{svplsSurr}
\description{
The \code{svplsSurr} class. An object of this class contains the following two slots:

surr A \code{data.frame} of the optimal surrogate variables.

prop.vars A vector of the variance proportions in the data space that are
	    explained by the variables in \code{surr}.	
}
