% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{svplsTest-class}
\alias{svplsTest-class}
\title{svplsTest}
\description{
The \code{svplsTest} class. An object of this class contains the following three slots:
	
pvs.unadj The uncorrected pvalues corresponding to the genes after 
adjusting for the signatures of hidden variability.

pvs.adj The multiple hypothesis corrected pvalues after adjusting
for the signatures of hidden variability.

sig.genes The genes detected to be significantly differentially 
expressed between the two groups.
}

