% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods.R
\docType{methods}
\name{pvs.adj}
\alias{pvs.adj}
\alias{pvs.adj,svplsTest-method}
\alias{pvs.adj.svplsTest}
\title{Accessor for the 'pvs.adj' slot of a 'svplsTest' object}
\usage{
pvs.adj(object)

\S4method{pvs.adj}{svplsTest}(object)

\S4method{pvs.adj}{svplsTest}(object)
}
\arguments{
\item{object}{a \code{svplsTest} object}
}
\value{
A vector of the FDR-adjusted pvalues for the different genes, 
obtained from \code{\link{svplsTest}}.
}
\description{
Accessor for the 'pvs.adj' slot of a 'svplsTest' object
}
\examples{
data(sim.dat)
group = as.factor(c(rep(1, 10), rep(-1, 10)))
sv = svplsSurr(sim.dat, group)
surr = surr(sv)
fit = svplsTest(dat = sim.dat, group = group, surr = surr, test = "Wald")
pvs.adj(fit)
}

