% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svplsSurr.R
\name{svplsSurr}
\alias{svplsSurr}
\title{svplsSurr}
\usage{
svplsSurr(dat, group, controls = NULL, phi = function(x) log(x + const),
  const = 1, pls.method = "oscorespls", max.surrs = 3, opt.surrs = 1,
  surr.select = c("automatic", "manual"), cutoff = 10^-7,
  parallel = FALSE, num.cores = NULL, plot = FALSE)
}
\arguments{
\item{dat}{The original feature expression count matrix.}

\item{group}{a factor representing the sample indices belonging to the two
different groups.}

\item{controls}{The set of control features with no differential expression
between the two groups (set to NULL by default).}

\item{phi}{The transforming function to be applied on the original feature
expression count data (set to be log function with an offset \code{const}).}

\item{const}{The offset parameter for the transforming function \code{phi}
(set to 1 by default).}

\item{pls.method}{The non-linear partial least squares method to be used.
The different options available are: the classical orthogonal scores
algorithm ("oscorespls, default), the kernel algorithm ("kernelpls") and
wide kernel algorithm ("widekernelpls"). Using the "oscorespls" option is
recommended for producing mutually orthogonal surrogate variables.}

\item{max.surrs}{The maximum number of factor estimates to be extracted
from the NPLS algorithm (set to 3 by default).}

\item{opt.surrs}{The index vector of factor estimates to be taken as the
optimal surrogate variables (used for manual selection only).}

\item{surr.select}{The method for selecting the optimal surrogate variables
("automatic" or "manual").}

\item{cutoff}{The user-specified pvalue cutoff for testing the significance
of the extracted surrogate variables (set to 1e-07 by default) (used for "automatic" selection only).}

\item{parallel}{Logical, indicating if the computations should be
parallelized or not (set to \code{FALSE} by default).}

\item{num.cores}{The requested number of cores to be used in the parallel
computations inside the function (used only when \code{parallel} is
\code{TRUE}, \code{NULL} by default).}

\item{plot}{Logical, if \code{TRUE} a barplot of the variance proportions
explained by the significant surrogate variables is returned (set to
\code{FALSE} by default).}
}
\value{
surr A \code{data.frame} of the optimal surrogate variables.

prop.vars A vector of the variance proportions explained by the
variables in \code{surr}.
}
\description{
This function extracts the surrogated estimates of the hidden variables in the
data by using the partial least squares (PLS) algorithm
on two multivariate random matrices. It provides the user with two options:

(1) \bold{Unsupervised SVAPLS}: Here a standard linear regression model is first used on
a transformed version of the expression count matrix to estimate the primary signals
of differential expression for all the features. The fitted model residuals and the
transformed count matrix are then organized respectively into two multivariate matrices
\code{E} and \code{Y}, in such a way that each column corresponds to a certain feature.
\code{Y} is then regressed on \code{E} using a Non-linear partial least squares (NPLS)
algorithm and the extracted factor estimates (scores) in the column-space of \code{Y}
are deemed as the surrogate variables.

(2) \bold{Supervised SVAPLS}: In case information on a set of control features (control genes, transcripts, spike-ins, etc.)
is provided, this function uses a Non-linear partial least squares (NPLS) algorithm to regress \code{Y} on another expression
matrix \code{Y.cont} corresponding to the set of controls and the factor estimates (scores) in the column-space of \code{Y.cont}
are considered as the surrogate variables.

An optimal subset of these variables is then selected either manually by the user (manual selection) or by testing them for
statistical significance (automatic selection). For the automatic selection the function  regresses the first right
singular vector of the residual matrix \code{E} (for Unsupervised SVAPLS) or the control matrix \code{Y.cont}
(for Supervised SVAPLS), on all the surrogate variables and the estimated regression coefficients are used to perform
a t-test with a certain user-specified pvalue cutoff. The variables yielding a pvalue below the cutoff are returned
as the optimal surrogate variables.
}
\examples{
##Loading a simulated RNAseq gene expression count dataset
data(sim.dat)

##Extracting the optimal surrogate variables
group = as.factor(c(rep(1, 10), rep(-1, 10)))
sv <- svplsSurr(dat = sim.dat, group = group, surr.select = "automatic")
slotNames(sv)
head(surr(sv))
head(prop.vars(sv))

}
